Locking an image managed by Marathon/Mesos
Sometimes a running container can crash and Marathon will automatically try to restart the container on a different node within the cluster. When this happens, the rbd-docker-plugin running on the other node will not be able to gain an exclusive lock on the image as it is still locked by the previous container that crashed.

As a workaround, the locker cookie is now generated through an md5sum of the pool + image name. By doing so the lock ID can always be generated by the rbd-docker-plugin running on any node in the cluster.